+++ packages/table/babel.config.cjs
@@ -1,5 +1,5 @@
-// Don't process node_modules except for AWS UI.
+// Don't process node_modules except for cloudscape-design.
 module.exports = {
-  ignore: [/node_modules\/?!@awsui\/components-react/],
+  ignore: [/node_modules\/?!@cloudscape-design\/components/],
   presets: ['@babel/preset-env'],
 };