+++ packages/table/src/table/index.tsx
@@ -1,6 +1,6 @@
 import React, { FunctionComponent } from 'react';
-import { PropertyFilter, Table as AWSUITable } from '@awsui/components-react';
-import { useCollection } from '@awsui/collection-hooks';
+import { PropertyFilter, Table as CSTable } from '@cloudscape-design/components';
+import { useCollection } from '@cloudscape-design/collection-hooks';
 import { TableProps } from '../utils';
 import { getDefaultColumnDefinitions } from '../utils/tableHelpers';
 
@@ -15,7 +15,7 @@ export const Table: FunctionComponent<TableProps> = (props) => {
   const { items, collectionProps, propertyFilterProps } = useCollection(userItems, { sorting, propertyFiltering });
   const columnDefinitions = getDefaultColumnDefinitions(userColumnDefinitions);
   return (
-    <AWSUITable
+    <CSTable
       {...props}
       items={items}
       {...collectionProps}